refactor: 패키지 구조 변경으로 포인트 도메인 모듈화
- 포인트 관련 클래스들을 domain, service, controller, repository 패키지로 분리하여
  모듈화함.
- 각 패키지에 따라 책임을 명확히 하고, 유지보수성을 향상시킴.
