--- src/main/java/io/hhplus/tdd/point/PointHistory.java
@@ -1,10 +1,0 @@
-package io.hhplus.tdd.point;
-
-public record PointHistory(
-        long id,
-        long userId,
-        long amount,
-        TransactionType type,
-        long updateMillis
-) {
-}
