+++ src/main/java/io/hhplus/tdd/point/service/PointService.java
@@ -1,0 +1,7 @@
+package io.hhplus.tdd.point.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class PointService {
+}
